added tar file heudiconv capability

+++ run.py
@@ -7,6 +7,7 @@ import argparse
 from IntendedFor import setup
 import sys
 import pdb
+import tarfile
 
 #book keeping and global variables
 parser = argparse.ArgumentParser(description='Script that controls BIDS conversion for individual studies')
@@ -40,11 +41,19 @@ else:
 if args.ses_id:
 	if len(glob(args.dicom_dir + "/*" + args.ses_id + "*")) == 1:
 		if args.subj_id:
-			shutil.copytree(glob(args.dicom_dir + "/*" + args.ses_id + "*")[0],"/temp_dicom_dir/" + args.subj_id +"/" + args.ses_id)
-			convert_format = '/neurodocker/startup.sh heudiconv "-d %s/{subject}/{session}/*/* -s %s -ss %s --overwrite -o %s/BIDS_output"' % ("/temp_dicom_dir",  args.subj_id, args.ses_id, args.output_dir)
+			#is it a tar file? If so convert tar file directly
+			if tarfile.is_tarfile(glob(args.dicom_dir + "/*" + args.ses_id + "*")[0]):
+				shutil.copytree(glob(args.dicom_dir + "/*" + args.ses_id + "*")[0],"/temp_dicom_dir/" + args.subj_id +"/" + args.ses_id)
+				convert_format = '/neurodocker/startup.sh heudiconv "-d %s/{subject}/{session} -s %s -ss %s --overwrite -o %s/BIDS_output"' % ("/temp_dicom_dir",  args.subj_id, args.ses_id, args.output_dir)
+			else:
+				shutil.copytree(glob(args.dicom_dir + "/*" + args.ses_id + "*")[0],"/temp_dicom_dir/" + args.subj_id +"/" + args.ses_id)
+				convert_format = '/neurodocker/startup.sh heudiconv "-d %s/{subject}/{session}/*/* -s %s -ss %s --overwrite -o %s/BIDS_output"' % ("/temp_dicom_dir",  args.subj_id, args.ses_id, args.output_dir)
 		else:
-			shutil.copytree(glob(args.dicom_dir + "/*" + args.ses_id + "*")[0],"/temp_dicom_dir/" + args.ses_id)
-			convert_format = '/neurodocker/startup.sh heudiconv "-d %s/{subject}/*/* -s %s --overwrite -o %s/BIDS_output"' % ("/temp_dicom_dir",  args.ses_id, args.output_dir)
+			if tarfile.is_tarfile(glob(args.dicom_dir + "/*" + args.ses_id + "*")[0]):
+				convert_format = '/neurodocker/startup.sh heudiconv "-d %s/*{subject}* -s %s --overwrite -o %s/BIDS_output"' % (args.dicom_dir, args.ses_id, args.output_dir)
+			else:
+				shutil.copytree(glob(args.dicom_dir + "/*" + args.ses_id + "*")[0],"/temp_dicom_dir/" + args.ses_id)
+				convert_format = '/neurodocker/startup.sh heudiconv "-d %s/{subject}/*/* -s %s --overwrite -o %s/BIDS_output"' % ("/temp_dicom_dir",  args.ses_id, args.output_dir)
 	elif len(glob(args.dicom_dir + "/*" + args.ses_id + "*")) > 1:
 		raise Exception("There are multiple directories within dicom directory: " + args.dicom_dir + " with the same session id: " + args.ses_id + ". Must exit.")
 	else:
@@ -52,8 +61,11 @@ if args.ses_id:
 else:
 	if args.subj_id:
 		if len(glob(args.dicom_dir + "/*" + args.subj_id + "*")) == 1:
-			shutil.copytree(glob(args.dicom_dir + "/*" + args.subj_id + "*")[0],"/temp_dicom_dir/" + args.subj_id)	
-			convert_format = '/neurodocker/startup.sh heudiconv "-d %s/{subject}/*/* -s %s --overwrite -o %s/BIDS_output"' % ("/temp_dicom_dir",  args.subj_id, args.output_dir)
+			if tarfile.is_tarfile(glob(args.dicom_dir + "/*" + args.subj_id + "*")[0]):
+				convert_format = '/neurodocker/startup.sh heudiconv "-d %s/*{subject}* -s %s --overwrite -o %s/BIDS_output"' % (args.dicom_dir,  args.subj_id, args.output_dir)
+			else:
+				shutil.copytree(glob(args.dicom_dir + "/*" + args.subj_id + "*")[0],"/temp_dicom_dir/" + args.subj_id)	
+				convert_format = '/neurodocker/startup.sh heudiconv "-d %s/{subject}/*/* -s %s --overwrite -o %s/BIDS_output"' % ("/temp_dicom_dir",  args.subj_id, args.output_dir)
 		elif len(glob(args.dicom_dir + "/*" + args.subj_id + "*")) > 1:
 			raise Exception("There are multiple directories within dicom directory: " + args.dicom_dir + " with the same subjct id: " + args.subj_id + ". Must exit.")
 		else:
@@ -65,9 +77,11 @@ os.system(convert_format + convert_type)
 
 #once conversion is finished delete data from temp_dicom_dir
 if args.subj_id:
-	shutil.rmtree("/temp_dicom_dir/" + args.subj_id)
+	if os.path.isdir("/temp_dicom_dir/" + args.subj_id):
+		shutil.rmtree("/temp_dicom_dir/" + args.subj_id)
 elif args.ses_id:
-	shutil.rmtree("/temp_dicom_dir/" + args.ses_id)
+	if os.path.isdir("/temp_dicom_dir/" + args.ses_id):
+		shutil.rmtree("/temp_dicom_dir/" + args.ses_id)
 
 
 #now change IntendedFor field within fmaps
